+++ src/InflateStream.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace GuzzleHttp\Psr7;
 
+use GuzzleHttp\Psr7\Exception\TimeoutException;
 use Psr\Http\Message\StreamInterface;
 
 /**
@@ -23,8 +24,11 @@ final class InflateStream implements StreamInterface
 
     private StreamInterface $stream;
 
+    private ?StreamInterface $source;
+
     public function __construct(StreamInterface $stream)
     {
+        $this->source = $stream;
         $resource = StreamWrapper::getResource($stream);
         // Specify window=15+32, so zlib will use header detection to both gzip (with header) and zlib data
         // See https://www.zlib.net/manual.html#Advanced definition of inflateInit2
@@ -32,5 +36,43 @@ final class InflateStream implements StreamInterface
         // Default window size is 15.
         stream_filter_append($resource, 'zlib.inflate', STREAM_FILTER_READ, ['window' => 15 + 32]);
         $this->stream = $stream->isSeekable() ? new Stream($resource) : new NoSeekStream(new Stream($resource));
+    }
+
+    public function read(int $length): string
+    {
+        if ($length <= 0 || $this->source === null) {
+            return $this->stream->read($length);
+        }
+
+        try {
+            $data = $this->stream->read($length);
+        } catch (TimeoutException $e) {
+            throw $e;
+        } catch (\RuntimeException $e) {
+            if (StreamTimeout::isReadTimedOut($this->source)) {
+                throw new TimeoutException('Unable to read from stream: timed out', 0, $e);
+            }
+
+            throw $e;
+        }
+
+        if ($data === '' && StreamTimeout::isReadTimedOut($this->source)) {
+            throw new TimeoutException('Unable to read from stream: timed out');
+        }
+
+        return $data;
+    }
+
+    public function close(): void
+    {
+        $this->source = null;
+        $this->stream->close();
+    }
+
+    public function detach()
+    {
+        $this->source = null;
+
+        return $this->stream->detach();
     }
 }